params retyping; polynomials and their lvl2 specialization

+++ src/params.rs
@@ -5,10 +5,10 @@ pub const R: i32 = 1753; //2Nth root of unity mod Q
 pub const D: i32 = 13; //dropped bits
 
 // Implementation specific values
-pub const SEEDBYTES: i32 = 32;
-pub const CRHBYTES: i32 = 64;
-pub const POLYT1_PACKEDBYTES: i32 = 320;
-pub const POLYT0_PACKEDBYTES: i32 = 416;
+pub const SEEDBYTES: usize = 32;
+pub const CRHBYTES: usize = 64;
+pub const POLYT1_PACKEDBYTES: usize = 320;
+pub const POLYT0_PACKEDBYTES: usize = 416;
 
 // Specific security levels parameters
 pub mod lvl2;
